lib: Rename two header files.
