+++ liblepton/include/liblepton/liblepton.h
@@ -33,8 +33,8 @@ G_BEGIN_DECLS
 
 #include <liblepton/color.h>
 #include <liblepton/fill_type.h>
-#include <liblepton/line_type.h>
-#include <liblepton/line_cap_type.h>
+#include <liblepton/stroke_type.h>
+#include <liblepton/stroke_cap_type.h>
 #include <liblepton/point.h>
 #include <liblepton/str.h>
