+++ liblepton/include/liblepton/stroke_cap_type.h
