+++ liblepton/include/liblepton/stroke_type.h
